Use long id's instead of int id's

+++ src/main/java/be/aca/devcon/jockey/service/api/JockeyService.java
@@ -20,7 +20,7 @@ public interface JockeyService {
 	 * Finds the jockey with the specified id.
 	 * Returns null if the jockey cannot be found.
 	 */
-	Jockey getJockey(int id);
+	Jockey getJockey(long id);
 
 	/**
 	 * Find all teams in the Guest site.
